+++ tests/cmake/AddTestMaps.cmake
@@ -186,6 +186,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1556_boolean_parameters_treat_nil_as_true"
   "bugs/1562_crash_entity_on_stream"
   "bugs/1565_state_set_can_control_direction"
+  "bugs/1583_crash_reset"
   "bugs/1585_get_entity_by_name"
   "bugs/1600_map_loading_single_cam/map1"
   "bugs/1609_crash_save_lua_keyword"