+++ changelog.md
@@ -8,6 +8,7 @@
 * Fix crash when a moving custom entity activates a stream (#1562).
 * Fix crash when a sprite is synchronized to a sprite with more frames (#1117).
 * Fix crash when entering a map with tileset-specific sprites nearby.
+* Fix crash when resetting the game from the update step (#1583).
 * Fix return of invalid sprite object with bad sprite id (#1423).
 * Fix separator region detection corner cases (#1425).
 * Fix error reporting when creating entities with a non-existent model (#1494).